Continuing refactoring.
